chore: fix panic in psed for tfgen (#860)
In an upcoming bridge release we have added a new Call() method on
explicit providers. This is somewhat new functionality and it defeats
the experimental code in datadog that is used to reduce accidental
recursive type explosion. Adding a quick workaround that permits the
code to succeed even with the new bridge.